+++ Doppler/Utils/DopplerTheme.cs
@@ -1,8 +1,0 @@
-﻿namespace Doppler.Utils
-{
-    public enum DopplerTheme
-    {
-        LIGHT,
-        DARK
-    }
-}