+++ Doppler/Utils/FormatConversion.cs
@@ -1,9 +1,0 @@
-﻿namespace Doppler.Utils
-{
-    public static class FormatConversion
-    {
-        public static string MP3 => "MP3";
-        public static string MP4 => "MP4";
-
-    }
-}